`manage-expensive` script enables or disables expensive resources in env
It depends on an `order.txt` file in each environment's directory to
define the order.

Was not able to use Terragrunt's run-all functionality because, when
destroying expensive resources, it would run from root to leaf which
is the wrong order. So if the dependency tree looks like
"base -> service", and service has LB listeners attached to an ALB
that would get destroyed in "base", then the destruction of resources
would fail because AWS would not let the ALB get destroyed.

By using our own method, we can reverse the order when destroying
expensive resources.